ver 1.1 (with fix for issue #1)
