+++ README.md
@@ -1,0 +1,32 @@
+# watir_test
+
+Watir based test suite for testing https://demoapp.strongqa.com/ using Watir
+
+## Getting Started
+
+To obtain the application:
+
+`$ git clone https://github.com/fordelive/watir_test.git`
+
+or just download it from GitHub
+
+To run it step into the 'watir_test' directory issue the command:
+
+`$ ruby test_suites/test_suite.rb`
+
+To abort execution press `Ctrl+C`
+
+The results of the tests execution can be found in `results.txt` which is created (recreated) during each program execution.
+
+
+### Prerequisites
+
+```
+Ruby 2.6.0
+watir 6.16.5
+webdrivers 3.7.2
+```
+
+### Installing
+
+Installation is not needed.
